# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Drei, a self-hosted Git repository platform",
        "title": "Drei API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:3200",
    "basePath": "/api",
    "paths": {
...
        "/issues": {
            "get": {
                "description": "Returns issues from all repositories with optional filters for state, author, assignee, search, label, and sort",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Issues"
                ],
                "summary": "List all issues across repositories",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Issue state filter (open or closed)",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Author user ID",
                        "name": "author",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Assignee user ID",
                        "name": "assignee",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term (matches title, description, number)",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Label name",
                        "name": "label",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order (newest, oldest, recently-updated, etc.)",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/user/appearance": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    },
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "Returns the authenticated user's theme and language preferences\nUpdates the authenticated user's theme and language preferences",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "tags": [
                    "Settings",
                    "Settings"
                ],
                "summary": "Update appearance settings",
                "parameters": [
                    {
                        "description": "Theme (light|dark|system) and language (en)",
                        "name": "appearance",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "SessionAuth": []
                    },
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "Returns the authenticated user's theme and language preferences\nUpdates the authenticated user's theme and language preferences",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "tags": [
                    "Settings",
                    "Settings"
                ],
                "summary": "Update appearance settings",
                "parameters": [
                    {
                        "description": "Theme (light|dark|system) and language (en)",
                        "name": "appearance",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "description": "Returns all registered users",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "List all users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "SessionAuth": {
            "description": "Better-auth session cookie for authenticated requests",
            "type": "apiKey",
            "name": "better-auth.session_token",
            "in": "cookie"
        }
    }
}